chore: Renamed file extension (v2 addon)
