+++ docs/sample-v2-addon/src/components/navigation-menu.module.css
