+++ docs/sample-v2-addon/src/components/ui/page.ts
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
 import { WithBoundArgs } from '@glint/template';
 
-import styles from './page.css';
+import styles from './page.module.css';
 import type UiPageDemoComponent from './page/demo';
 import type UiPageSectionComponent from './page/section';
 import type UiPageSubsectionComponent from './page/subsection';
