+++ docs/sample-v2-addon/src/components/ui/page/subsection.ts
@@ -1,6 +1,6 @@
 import Component from '@glimmer/component';
 
-import styles from './subsection.css';
+import styles from './subsection.module.css';
 
 interface UiPageSubsectionSignature {
   Blocks: {
